feat: flesh out the vega render lifecycle
